+++ packages/tests-render-mithril/src/page.js
@@ -4,7 +4,7 @@ import Footer from "./Footer";
 
 const iconBack = h.trust("<svg width=\"24\" height=\"24\" viewBox=\"0 0 24 24\"><path d=\"M20 11H7.83l5.59-5.59L12 4l-8 8 8 8 1.41-1.41L7.83 13H20v-2z\"/></svg>");
 
-const navBar = (name, previous) =>
+const navBar = (name, previous, doc) =>
   h(css.headerRow, h(Toolbar, {
     style: { backgroundColor: "rgba(255,255,255,.93)" }
   }, [
@@ -16,10 +16,11 @@ const navBar = (name, previous) =>
       },
       style: { color: "#0091EA" }
     }),
-    h("span", name)
+    h("span", name),
+    doc && h(`a[href=${doc}][target=_blank][rel="noopener noreferrer"].pe-action`, "doc")
   ]));
 
-const results = (name, tests) => (
+const results = ({ name, tests }) => (
   h([css.results].join(" "), {
     className: `tests-${name.replace(/[^\w\d]+/g, "-").toLowerCase()}`
   }, tests.map(test => {
@@ -44,13 +45,13 @@ const results = (name, tests) => (
   }))
 );
   
-export default (name, tests, previous) => ({
+export default (route, previous) => ({
   oncreate: () => ( 
     scrollTo(0, 0)
   ),
   view: () => [
-    navBar(name, previous),
-    results(name, tests),
+    navBar(route.name, previous, route.doc),
+    results(route),
     h(Footer),
     h(Dialog),
     h(Snackbar),